feat(psychometric): V7 schema — decimal STEN, T-score, composite/strategy columns
Flyway V7 migration adds: decimal STEN (NUMERIC 4,2) on scale_score and
norm_entry (with matview drop/recreate and chk_sten constraint refresh),
T-score column on scale_score, T-score params (tFactor/tOffset/tClipLo/tClipHi)
on norm_scale_param, composite_method/composite_basis on psychometric_scale,
norm_strategy on norm_table_version, item_strategy on scoring_key_item, and
result_state/validity_status on test_result. Entity fields, six new enums, and
ScaleScoreDto/NormEntryDto/NormEntryRequest updated to match; all call sites
in ScoringService and CandidatePsychometricService fixed to compile cleanly.

+++ src/main/java/com/edge/pulse/data/dto/psychometric/NormEntryDto.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 public record NormEntryDto(
         UUID scaleId,
         String scaleName,
-        int stenScore,
+        BigDecimal stenScore,
         BigDecimal rawScoreMin,
         BigDecimal rawScoreMax,
         BigDecimal percentile,
